Preserved ebuild for Apache vlogger.
The vlogger script for the efficient handling of Apache log files in a
virtual hosting environment was once available as a Gentoo package. As
the original script was no longer maintained, the ebuild was removed by
the Gentoo team. This is a modified version of the original ebuild, used
to install the fork of vlogger maintained by i-MSCP:
https://github.com/i-MSCP/imscp/tree/1.5.x/configs/debian/apache